pyproject.toml
[tool.poetry]
name = "robust_detection"
version = "0.1.0"
description = ""
authors = ["Anonymous <>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torch = "^1.10.2"
torchvision = "^0.11.3"
pyswip = "^0.2.10"
opencv-python = "^4.5.5"
tqdm = "^4.62.3"
pytorch-lightning = "^1.5.9"
scipy = "^1.7.3"
scikit-image = "^0.19.1"
pandas = "^1.4.0"
rdkit-pypi = "^2021.9.4"
ipdb = "^0.13.9"
wandb = "^0.12.10"
jupyter = "^1.0.0"
scikit-learn = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
